LPS-166857 Rename, follow CreatorUtil pattern
